+++ app/workers/download_attachment_worker.rb
@@ -6,5 +6,9 @@ class DownloadAttachmentWorker
   def perform(attachment_id, url, uploaded_by_user_id)
     user = User.find(uploaded_by_user_id)
     Attachment.find(attachment_id).download!(url, uploaded_by: user)
+
+  rescue ActiveRecord::RecordNotFound => ex
+    Rails.logger.info "Caught Attachment cancel: #{ex.message}"
+    # No-op. This is a user canceling a processing job
   end
 end