+++ client/app/models/paper.js
@@ -26,7 +26,9 @@ export default DS.Model.extend({
   groupAuthors: hasMany('group-author', { async: false }),
   journal: belongsTo('journal', { async: true }),
   manuscriptPageTasks: hasMany('task', { async: true, polymorphic: true }),
+
   paperTaskTypes: hasMany('paper-task-type', { async: true }),
+
   phases: hasMany('phase', { async: true }),
   relatedArticles: hasMany('related-article', { async: true }),
   snapshots: hasMany('snapshot', { inverse: 'paper', async: true }),
@@ -80,8 +82,8 @@ export default DS.Model.extend({
   submissionTasks: computed.filterBy('tasks', 'isSubmissionTask', true),
   sortedSubmissionTasks: computed.sort('submissionTasks', 'taskSorting'),
 
-  displayTitle: computed('title', 'shortTitle', function() {
-    return this.get('title') || this.get('shortTitle');
+  displayTitle: computed('title', function() {
+    return this.get('title') || '[No Title]';
   }),
 
   collaborators: computed('collaborations.@each.user', function() {
@@ -100,7 +102,8 @@ export default DS.Model.extend({
     'decisions.@each.latestRegistered',
     function() {
       return this.get('decisions').findBy('latestRegistered', true);
-  }),
+    }
+  ),
 
   previousDecisions: computed('decisions.@each.registeredAt', function() {
     return this.get('decisions')
@@ -234,7 +237,7 @@ export default DS.Model.extend({
       // the records which were pushed.
       this.store.pushPayload(data);
       return _.map(data['users'], (user) => {
-        return this.store.find('user', user['id']);
+        return this.store.findRecord('user', user['id']);
       });
     });
   }