+++ client/app/pods/components/attachment-manager/component.js
@@ -27,14 +27,15 @@ export default Ember.Component.extend({
   disabled: false,
   notDisabled: Ember.computed.not('disabled'),
   buttonText: 'Upload File',
-  fileUploads: [],
+  fileUploads: Ember.computed(() => { return []; }),
   multiple: false,
   showDescription: true,
 
   uploadInProgress: Ember.computed.notEmpty('fileUploads'),
-  hasAttachments: Ember.computed.notEmpty('attachments'),
-  noAttachments: Ember.computed.not('hasAttachments'),
-  canUploadFile: Ember.computed.or('noAttachments', 'multiple'),
+  canUploadFile: Ember.computed('attachments.[]', 'multiple', function() {
+    return Ember.isEmpty(this.get('attachments')) || this.get('multiple');
+  }),
+
   showAddButton: Ember.computed.and('notDisabled', 'canUploadFile'),
 
   init() {
@@ -72,7 +73,7 @@ export default Ember.Component.extend({
     },
 
     uploadFailed(reason){
-     console.log('uploadFailed', reason);
+      throw new Ember.Error(`s3 uploadFailed: ${reason}`);
     }
   }
 });