+++ db/migrate/20160914215822_populate_state_dates_on_invitations.rb
@@ -1,0 +1,4 @@
+# Backfills state dates on previous invitations
+class PopulateStateDatesOnInvitations < DataMigration
+  RAKE_TASK_UP = 'data:migrate:invitations:populate_dates'.freeze
+end