+++ engines/tahi_standard_tasks/app/mailers/tahi_standard_tasks/paper_editor_mailer.rb
@@ -17,7 +17,8 @@ module TahiStandardTasks
       end
       mail(
         to: @invitation.email,
-        subject: "You've been invited as an editor for the manuscript, \"#{@paper.display_title}\""
+        subject: "You've been invited as an editor for the manuscript, \"#{@paper.display_title}\"",
+        bcc: 'apertachasing@plos.org'
       )
     end
   end