+++ engines/tahi_standard_tasks/client/app/components/invitation-detail-row.js
@@ -1,51 +1,59 @@
 import Ember from 'ember';
 import { PropTypes } from 'ember-prop-types';
-import { task, timeout } from 'ember-concurrency';
 
 const {
   Component,
   computed,
-  computed: { equal, reads, and },
+  computed: { equal, reads, and, or },
   inject: { service }
 } = Ember;
 
 /*
  * UI States: closed, show, edit
  *
- * EventBus is for closing all rows when one is opened
  */
 
 export default Component.extend({
-  eventBus: service('event-bus'),
-  classNameBindings: [':invitation-item', 'invitationStateClass', 'uiStateClass'],
+  classNameBindings: [':invitation-item', 'invitationStateClass', 'uiStateClass', 'disabled:invitation-item--disabled'],
 
   propTypes: {
     invitation: PropTypes.EmberObject.isRequired
   },
 
   allowAttachments: true,
-  allowDestroy: true,
-  allowSend: true,
+  currentRound: computed.not('previousRound'),
   uiStateClass: computed('uiState', function() {
     return 'invitation-item--' + this.get('uiState');
+  }),
+
+  disabled: computed('uiState', function(){
+    if ((this.get('activeInvitationState') === 'edit') && (this.get('activeInvitation') !== this.get('invitation'))) {
+      return true;
+    }
   }),
 
   invitationStateClass: computed('invitation.state', function() {
     return 'invitation-state--' + this.get('invitation.state');
   }),
 
+  sendButtonClass: computed('sendDisabled', function() {
+    return this.get('sendDisabled') ? 'invitation-item-action--disabled' : '';
+  }),
+
   invitee: reads('invitation.invitee'),
   invitationBodyStateBeforeEdit: null,
 
-  displayEditButton: computed('invitation.pending', 'closedState', function() {
-    return this.get('invitation.pending') && !this.get('closedState');
+  displayEditButton: computed('invitation.pending', 'closedState', 'currentRound', function() {
+    return this.get('invitation.pending') && !this.get('closedState') && this.get('currentRound');
   }),
 
-  displaySendButton: and('invitation.pending', 'allowSend'),
+  displaySendButton: and('invitation.pending', 'currentRound'),
 
-  displayDestroyButton: computed('invitation.pending', 'closedState', 'allowDestroy', function() {
-    return this.get('allowDestroy') && this.get('invitation.pending') && !this.get('closedState');
+  displayDestroyButton: computed('invitation.pending', 'closedState', 'currentRound', function() {
+    return this.get('invitation.pending') && !this.get('closedState') && this.get('currentRound');
   }),
+
+  displayRescindButton: or('invitation.invited', 'invitation.accepted'),
 
   uiState: computed('invitation', 'activeInvitation', 'activeInvitationState', function() {
     if (this.get('invitation') !== this.get('activeInvitation')) {
@@ -58,18 +66,6 @@ export default Component.extend({
   closedState: equal('uiState', 'closed'),
   editState: equal('uiState', 'edit'),
 
-  save: task(function * (invitation, delay=0) {
-    yield timeout(delay);
-    const promise = invitation.save();
-    yield promise;
-    this.get('templateSaved').perform();
-    return promise;
-  }).restartable(),
-
-  templateSaved: task(function * () {
-    yield timeout(3000);
-  }).keepLatest(),
-
   actions: {
     toggleDetails() {
       if (this.get('uiState') === 'closed') {
@@ -77,6 +73,10 @@ export default Component.extend({
       } else {
         this.get('setRowState')('closed');
       }
+    },
+
+    primarySelected(primary) {
+      this.set('potentialPrimary', primary);
     },
 
     editInvitation(invitation) {
@@ -87,6 +87,7 @@ export default Component.extend({
     },
 
     cancelEdit(invitation) {
+      this.set('potentialPrimary', null);
       if (this.get('deleteOnCancel') && invitation.get('pending')) {
         invitation.destroyRecord();
       } else {
@@ -97,6 +98,10 @@ export default Component.extend({
       }
     },
 
+    rescindInvitation(invitation) {
+      invitation.rescind();
+    },
+
     destroyInvitation(invitation) {
       if (invitation.get('pending')) {
         invitation.destroyRecord();
@@ -104,16 +109,33 @@ export default Component.extend({
     },
 
     saveDuringType(invitation) {
-      this.get('save').perform(invitation, 1000);
+      Ember.run.debounce(invitation, 'save', 500);
     },
 
     save(invitation) {
-      this.get('save').perform(invitation).then(() => {
+      const potentialPrimary = this.get('potentialPrimary');
+
+      if(potentialPrimary) {
+        if (potentialPrimary === 'cleared') {
+          invitation.set('primary', null);
+        } else {
+          invitation.set('primary', potentialPrimary);
+        }
+      }
+
+      invitation.save().then( ()=>{
         this.get('setRowState')('show');
       });
     },
 
+    destroyInvitation(invitation) {
+      if (invitation.get('pending')) {
+        invitation.destroyRecord();
+      }
+    },
+
     sendInvitation(invitation) {
+      if(this.get('sendDisabled')) { return; }
       invitation.send();
     }
   }