+++ spec/services/journal_factory_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe JournalFactory do
+describe JournalFactory, flaky: true do
   describe '.create' do
     include AuthorizationSpecHelper
 
@@ -51,6 +51,30 @@ describe JournalFactory do
         journal = JournalFactory.create(name: 'Journal of the Stars')
         expect(journal.name).to eq('Journal of the Stars')
       end.to change(Journal, :count).by(1)
+    end
+
+    context 'role hints' do
+      let!(:journal) { JournalFactory.create(name: 'Journal of the Stars') }
+
+      it 'assigns hints to discussion topic roles' do
+        expect(Role::DISCUSSION_TOPIC_ROLES.sort).to \
+          eq(journal.roles.where(assigned_to_type_hint: DiscussionTopic.name).map(&:name))
+      end
+
+      it 'assigns hints to task roles' do
+        expect(journal.roles.where(assigned_to_type_hint: Task.name).map(&:name).sort)
+          .to eq(Role::TASK_ROLES.sort)
+      end
+
+      it 'assigns paper hints to paper roles' do
+        expect(journal.roles.where(assigned_to_type_hint: Paper.name).map(&:name).sort)
+          .to eq(Role::PAPER_ROLES.sort)
+      end
+
+      it 'assigns journal hints to journal roles' do
+        expect(journal.roles.where(assigned_to_type_hint: Journal.name).map(&:name).sort)
+          .to eq(Role::JOURNAL_ROLES.sort)
+      end
     end
 
     context 'creating the default roles and permission for the journal', flaky: true do
@@ -1050,10 +1074,9 @@ describe JournalFactory do
 
         describe 'permission to PlosBilling::BillingTask' do
           it 'can :view and :edit' do
-            expect(permissions).to include(
-              Permission.find_by(action: 'view', applies_to: 'PlosBilling::BillingTask'),
-              Permission.find_by(action: 'edit', applies_to: 'PlosBilling::BillingTask')
-            )
+            # Sometimes there is more than one 'edit' or 'view' permission for BillingTask so this fixes spec flakiness
+            permission_strings = permissions.where(applies_to: 'PlosBilling::BillingTask').pluck(:action)
+            expect(permission_strings).to contain_exactly('view', 'edit')
           end
         end
       end